driver: crypto: qce: Fix warning in qce driver
This change will fix warning in qce driver.

warning:
[comparison of distinct pointer
types lacks a cast
  (void) (&_max1 == &_max2);  \

The type of second argument of max macro is int only.
so typcast second argument of max macro to int.

+++ drivers/crypto/qce/ablkcipher.c
@@ -158,7 +158,7 @@ static int qce_setkey_sec(unsigned int keylen)
 	int ret;
 
 	qce_cache = kmem_cache_create("qce_sec_key", max(L1_CACHE_BYTES,
-				sizeof(struct qce_config_key_sec)),
+				(int)sizeof(struct qce_config_key_sec)),
 			L1_CACHE_BYTES, SLAB_PANIC, NULL);
 	if (!qce_cache) {
 		pr_info("%s: Error in creating cache\n",__func__);
